% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ajr}
\alias{ajr}
\title{Cross-national data on economic growth}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 163 rows and 15 columns.
}
\source{
Acemoglu, Johnson, and Robinson (2001)
}
\usage{
ajr
}
\description{
Data on the long-run effect of colonial institutions on economic
growth
}
\keyword{datasets}
