% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes}
\alias{anes}
\title{Selected variables from ANES 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5162 rows and 8 columns.
}
\source{
https://electionstudies.org
}
\usage{
anes
}
\description{
Selected variables from the American National Election Study (2020)
}
\keyword{datasets}
