% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cces_2020}
\alias{cces_2020}
\title{Selected variables from CES 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51551 rows and 6 columns.
}
\source{
Ansolabehere, Schaffner, and Luks (2021).
}
\usage{
cces_2020
}
\description{
Selected variables from the Congressional Election Study (2020)
}
\keyword{datasets}
