% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_years}
\alias{class_years}
\title{Data on the distribution of college years in an undergraduate class}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 122 rows and 1 columns.
}
\source{
Personal correspondence
}
\usage{
class_years
}
\description{
One column data frame with the year of college or graduate school
for a class
}
\keyword{datasets}
