% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health}
\alias{health}
\title{Data from a person's daily measurements of activity and weight}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 644 rows and 6 columns.
}
\source{
Personal correspondence
}
\usage{
health
}
\description{
Measurements of a person's weight and activity collected via
internet connected devices.
}
\keyword{datasets}
