% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midterms}
\alias{midterms}
\title{Election prediction data for US midterms}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 6 columns.
}
\source{
Varied
}
\usage{
midterms
}
\description{
Data on change in seats for the president's party, presdential
approval, and percentage change in real disposable income
}
\keyword{datasets}
