% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Under-5 child mortality data, World Bank}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 217 rows and 52 columns.
}
\source{
https://data.worldbank.org/indicator/SH.DYN.MORT
}
\usage{
mortality
}
\description{
Data on under 5 child mortality per 1,000 live births
}
\keyword{datasets}
