% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{news}
\alias{news}
\title{Local TV news data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3137 rows and 10 columns.
}
\source{
Martin and McCrain (2019).
}
\usage{
news
}
\description{
Data on content and slant of local TV news in 2017
}
\keyword{datasets}
