% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polls20}
\alias{polls20}
\title{Presidential polling data, 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3058 rows and 5 columns.
}
\source{
https://github.com/fivethirtyeight/data/tree/master/polls
}
\usage{
polls20
}
\description{
Data from 528 on state-level polling for Joe Biden and Donald Trump
in the 2020 US presidential election
}
\keyword{datasets}
