% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pres20}
\alias{pres20}
\title{Presidential election results, 2020}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51 rows and 5 columns.
}
\source{
https://electionlab.mit.edu/data
}
\usage{
pres20
}
\description{
Election results from the 2020 US presidential election at the
state level
}
\keyword{datasets}
