% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spotify}
\alias{spotify}
\title{Spotify US ranking data, 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 490 rows and 54 columns.
}
\source{
https://www.kaggle.com/datasets/yujisato/spotify-us-weekly-top200
}
\usage{
spotify
}
\description{
Ranks of 50 most popular songs streamd by Spotify users in the US,
during 2020
}
\keyword{datasets}
