% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tea}
\alias{tea}
\title{Lady Tasting Tea data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.
}
\source{
Fisher (1935).
}
\usage{
tea
}
\description{
Tibble of randomly assorted cups of tea either poured milk first or
tea first with the Lady's guesses from Fisher's Lady Tasting Tea example
}
\keyword{datasets}
