% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trains}
\alias{trains}
\title{Boston commuter rail outgroup exposure study}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 115 rows and 14 columns.
}
\source{
Enos (2014)
}
\usage{
trains
}
\description{
Data on an experiment exposing commuter rail riders to outgroup members
}
\keyword{datasets}
