% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans}
\alias{trans}
\title{Transphobia study data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 565 rows and 9 columns.
}
\source{
Broockman and Kalla (2016).
}
\usage{
trans
}
\description{
Data on an experiment to reduce anti-transgender prejuidice
}
\keyword{datasets}
