% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polls20}
\alias{polls20}
\alias{national_polls20}
\title{Presidential polling data, 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2445 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 654 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2445 rows and 7 columns.
}
\source{
https://github.com/fivethirtyeight/data/tree/master/polls

https://github.com/fivethirtyeight/data/tree/master/polls
}
\usage{
polls20

national_polls20

polls20
}
\description{
Data from 528 on state-level polling for Joe Biden and Donald Trump
in the 2020 US presidential election

Data from 528 on national and state-level polling for Joe Biden and
Donald Trump in the 2020 US presidential election
}
\keyword{datasets}
